revamped the implementation
The coordinator now uses the Queue class from the standard library to handle the job_queue. This simplified the worker a great deal -- which is now just a single line inside the Coordinator. The Coordinator also got a shutdown method which terminates the workers and clears the job_queue.

Increased test coverage and provided logic for edge cases in certain scenarios.